Organize publisher vault and render modules
