+++ crates/publish/src/render.rs
@@ -1,213 +1,10 @@
-use crate::BookmarkEnricher;
-use crate::artifacts::CategoryLandingMetadata;
-use crate::classify::{ParsedArticleFile, ParsedCategoryFile, ParsedHomeFile, ParsedPageFile};
-use crate::error::Result;
-use crate::ingest::convert_markdown_to_html;
-use crate::links;
-use domain::{
-    ArticleBody, ArticleMeta, ArticleMetaInput, Category, HomeFragmentArtifactDocument,
-    PageArtifactDocument, Title,
-};
-use log::warn;
+mod bookmark;
+mod content;
+mod html;
 
-pub(crate) struct RenderedArticle {
-    pub(crate) meta: ArticleMeta,
-    pub(crate) html: String,
-}
-
-pub(crate) struct RenderedPage {
-    pub(crate) document: PageArtifactDocument,
-}
-
-pub(crate) struct RenderedCategoryLanding {
-    pub(crate) metadata: CategoryLandingMetadata,
-    pub(crate) html: String,
-}
-
-async fn render_html(
-    markdown_body: &str,
-    link_index: &links::Index,
-    enrich: &BookmarkEnricher,
-) -> Result<String> {
-    let markdown_with_links = links::convert(markdown_body, link_index);
-    let html_body = convert_markdown_to_html(&markdown_with_links)?;
-    let fallback = html_body.clone();
-    let html = enrich(html_body).await.unwrap_or_else(|e| {
-        warn!("Warning: Failed to convert simple bookmarks to rich bookmarks: {e}");
-        fallback
-    });
-    Ok(html)
-}
-
-pub(crate) async fn render_article(
-    parsed_file: ParsedArticleFile,
-    link_index: &links::Index,
-    enrich: BookmarkEnricher,
-) -> Result<RenderedArticle> {
-    let html = render_html(&parsed_file.markdown_body, link_index, &enrich).await?;
-    let meta = ArticleMeta::new(ArticleMetaInput {
-        slug: parsed_file.slug,
-        title: Title::new(parsed_file.front_matter.title)?,
-        category: parsed_file.category,
-        section_path: parsed_file.section_path,
-        description: parsed_file.front_matter.summary,
-        tags: parsed_file.front_matter.tags.unwrap_or_default(),
-        priority: parsed_file.front_matter.priority,
-        created_at: parsed_file.front_matter.created,
-        updated_at: parsed_file.front_matter.updated,
-    })?;
-    let body = ArticleBody::new(html)?;
-    Ok(RenderedArticle {
-        meta,
-        html: body.html,
-    })
-}
-
-pub(crate) async fn render_page(
-    parsed_file: ParsedPageFile,
-    link_index: &links::Index,
-    enrich: BookmarkEnricher,
-) -> Result<RenderedPage> {
-    let html = render_html(&parsed_file.markdown_body, link_index, &enrich).await?;
-    Ok(RenderedPage {
-        document: PageArtifactDocument {
-            page: parsed_file.page,
-            title: parsed_file.front_matter.title,
-            description: parsed_file.front_matter.summary,
-            html,
-            updated_at: parsed_file.front_matter.updated,
-        },
-    })
-}
-
-pub(crate) async fn render_home(
-    parsed_file: ParsedHomeFile,
-    link_index: &links::Index,
-    enrich: BookmarkEnricher,
-) -> Result<HomeFragmentArtifactDocument> {
-    let html = render_html(&parsed_file.markdown_body, link_index, &enrich).await?;
-    Ok(HomeFragmentArtifactDocument {
-        title: parsed_file.front_matter.title,
-        description: parsed_file.front_matter.summary,
-        html,
-        updated_at: parsed_file.front_matter.updated,
-    })
-}
-
-pub(crate) async fn render_category(
-    parsed_file: ParsedCategoryFile,
-    link_index: &links::Index,
-    enrich: BookmarkEnricher,
-) -> Result<RenderedCategoryLanding> {
-    let html = render_html(&parsed_file.markdown_body, link_index, &enrich).await?;
-    let title = normalize_category_title(parsed_file.category, &parsed_file.front_matter.title);
-    let description = normalize_category_description(parsed_file.front_matter.summary.as_deref());
-    let html = if html.trim().is_empty() {
-        build_fallback_category_landing_html(parsed_file.category, &title, description.as_deref())
-    } else {
-        html
-    };
-    Ok(RenderedCategoryLanding {
-        metadata: CategoryLandingMetadata {
-            category: parsed_file.category,
-            title,
-            description,
-            updated_at: parsed_file.front_matter.updated,
-        },
-        html,
-    })
-}
-
-fn normalize_category_title(category: Category, title: &str) -> String {
-    let title = title.trim();
-    if title.is_empty() {
-        category.display_name().to_owned()
-    } else {
-        title.to_owned()
-    }
-}
-
-fn normalize_category_description(description: Option<&str>) -> Option<String> {
-    description
-        .map(str::trim)
-        .filter(|description| !description.is_empty())
-        .map(str::to_owned)
-}
-
-fn build_fallback_category_landing_html(
-    category: Category,
-    title: &str,
-    description: Option<&str>,
-) -> String {
-    let heading = if title.trim().is_empty() {
-        category.display_name()
-    } else {
-        title.trim()
-    };
-
-    let body = description
-        .filter(|d| !d.trim().is_empty())
-        .map(str::trim)
-        .map(str::to_owned)
-        .unwrap_or_else(|| format!("{}カテゴリの記事一覧です。", category.display_name()));
-    let heading = html_escape(heading);
-    let body = html_escape(&body);
-    format!("<article><h1>{heading}</h1><p>{body}</p></article>")
-}
-
-fn html_escape(text: &str) -> String {
-    text.replace('&', "&amp;")
-        .replace('<', "&lt;")
-        .replace('>', "&gt;")
-        .replace('"', "&quot;")
-        .replace('\'', "&#39;")
-}
+pub use bookmark::BookmarkEnricher;
+pub(crate) use bookmark::rich_bookmark_enricher;
+pub(crate) use content::{render_article, render_category, render_home, render_page};
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_build_fallback_category_landing_html_uses_title_and_summary() {
-        let html = build_fallback_category_landing_html(
-            Category::Tech,
-            "Tech",
-            Some("Technology landing"),
-        );
-
-        assert!(html.contains("<h1>Tech</h1>"));
-        assert!(html.contains("<p>Technology landing</p>"));
-    }
-
-    #[test]
-    fn test_build_fallback_category_landing_html_falls_back_to_category_display_name() {
-        let html = build_fallback_category_landing_html(Category::Physics, "   ", None);
-
-        assert!(html.contains("<h1>物理学</h1>"));
-        assert!(html.contains("物理学カテゴリの記事一覧です。"));
-    }
-
-    #[test]
-    fn test_build_fallback_category_landing_html_escapes_frontmatter_text() {
-        let html = build_fallback_category_landing_html(
-            Category::Tech,
-            "<script>alert(1)</script>",
-            Some("\"quoted\" & <tag>"),
-        );
-
-        assert!(html.contains("&lt;script&gt;alert(1)&lt;/script&gt;"));
-        assert!(html.contains("&quot;quoted&quot; &amp; &lt;tag&gt;"));
-        assert!(!html.contains("<script>alert(1)</script>"));
-    }
-
-    #[test]
-    fn test_normalize_category_metadata_trims_and_falls_back() {
-        assert_eq!(normalize_category_title(Category::Physics, "   "), "物理学");
-        assert_eq!(normalize_category_title(Category::Tech, "  Tech  "), "Tech");
-        assert_eq!(normalize_category_description(Some("   ")), None);
-        assert_eq!(
-            normalize_category_description(Some("  Technology landing  ")),
-            Some("Technology landing".to_string())
-        );
-    }
-}
+pub(crate) use html::convert_markdown_to_html;
