+++ crates/publish/src/vault.rs
@@ -1,9 +1,7 @@
-mod converter;
 mod parser;
 mod scanner;
 
-pub(crate) use converter::convert_markdown_to_html;
 pub(crate) use parser::{
     ContentKind, ObsidianFrontMatter, ParsedObsidianFile, parse_obsidian_file,
 };
-pub(crate) use scanner::scan_obsidian_files;
+pub(crate) use scanner::scan_markdown_files;
