+++ crates/publish/src/vault/scanner.rs
@@ -3,8 +3,8 @@ use ignore::WalkBuilder;
 use std::path::{Path, PathBuf};
 
 /// Scans the specified directory for Markdown files (.md) and returns their paths.
-pub(crate) fn scan_obsidian_files(publish_dir: impl AsRef<Path>) -> Result<Vec<PathBuf>> {
-    let mut md_files: Vec<PathBuf> = WalkBuilder::new(publish_dir.as_ref())
+pub(crate) fn scan_markdown_files(vault_dir: impl AsRef<Path>) -> Result<Vec<PathBuf>> {
+    let mut md_files: Vec<PathBuf> = WalkBuilder::new(vault_dir.as_ref())
         .hidden(true)
         .git_ignore(false)
         .build()
@@ -55,7 +55,7 @@ mod tests {
             fs::write(base_path.join(file_path), "Other content")?;
         }
 
-        let files = scan_obsidian_files(base_path)?;
+        let files = scan_markdown_files(base_path)?;
         assert_eq!(files.len(), expected_count);
 
         Ok(())
@@ -75,7 +75,7 @@ mod tests {
         fs::write(base_path.join(filename), content)?;
         fs::write(base_path.join("not_markdown.txt"), "Not markdown")?;
 
-        let files = scan_obsidian_files(base_path)?;
+        let files = scan_markdown_files(base_path)?;
 
         assert_eq!(files.len(), 1);
         assert_eq!(files[0].file_name().unwrap().to_string_lossy(), filename);
